chore push things

+++ Test/Passes/isel-riscv.mlir
@@ -2,14 +2,14 @@
 
 "builtin.module"() ({
 ^bb0():
-    %one = "llvm.constant"() <{ "value" = 1 : i32 }> : () -> i32
-    %two = "llvm.constant"() <{ "value" = 2 : i32 }> : () -> i32
+    %one = "llvm.constant"() <{ "value" = 1 : i64 }> : () -> i64
+    %two = "llvm.constant"() <{ "value" = 2 : i64 }> : () -> i64
     
-    // CHECK:      %{{.*}} = "llvm.constant"() <{"value" = 1 : i32}> : () -> i32
-    // CHECK-NEXT: %{{.*}} = "llvm.constant"() <{"value" = 2 : i32}> : () -> i32
+    // CHECK:      %{{.*}} = "llvm.constant"() <{"value" = 1 : i64}> : () -> i64
+    // CHECK-NEXT: %{{.*}} = "llvm.constant"() <{"value" = 2 : i64}> : () -> i64
     
-    %add = "llvm.add"(%two, %one) : (i32, i32) -> i32
+    %add = "llvm.add"(%two, %one) : (i64, i64) -> i64
 
-    // CHECK-NEXT: %{{.*}} = "riscv.add"(%{{.*}}, %{{.*}}) : (i32, i32) -> i32
+    // CHECK-NEXT: %{{.*}} = "riscv.add"(%{{.*}}, %{{.*}}) : (i64, i64) -> i64
 
 }) : () -> ()
